requirements.txt
certifi==2019.3.9
cfenv==0.5.3
chardet==3.0.4
Click==7.0
Flask==1.0.2
furl==2.0.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
orderedmultidict==1.0
pycrypto==2.6.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.2
